Move paging logic into the matcher, and extract the logic into a function
