+++ internal/clients/clientimpl/osvmatcher/errors.go
@@ -1,4 +1,4 @@
-package osvdev
+package osvmatcher
 
 import "fmt"
